Isolate the event handlers to their own file
